requirements.txt
Markdown==3.5.1
MarkupSafe==2.1.3
Pillow==10.1.0
absl-py==2.0.0
astunparse==1.6.3
cachetools==5.3.2
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
et-xmlfile==1.1.0
flatbuffers==23.5.26
fonttools==4.46.0
gast==0.5.4
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.0
h5py==3.10.0
idna==3.6
imageio==2.33.1
importlib-metadata==7.0.0
importlib-resources==6.1.1
joblib==1.3.2
keras==2.15.0
kiwisolver==1.4.5
lazy-loader==0.3
libclang==16.0.6
matplotlib==3.8.2
mealpy==2.5.4
ml-dtypes==0.2.0
networkx==3.2.1
nltk==3.8.1
numpy==	1.26.2
oauthlib==	3.2.2
opencv-python==	4.8.1.78
openpyxl==	3.1.2
opfunu==	1.0.1
opt-einsum==	3.3.0
packaging==	23.2
pandas==	2.1.4
pip==	23.3.1
protobuf==	4.23.4
psutil==	5.9.6
pyasn1==	0.5.1
pyasn1-modules==	0.3.0
pyparsing==	3.1.1
python-dateutil==	2.8.2
pytz==	2023.3.post1
regex==	2023.10.3
requests==	2.31.0
requests-oauthlib==	1.3.1
rsa==	4.9
scikit-image==	0.22.0
scikit-learn==	1.3.2
scipy==	1.11.4
seaborn==	0.13.0
setuptools==	69.0.2
six==	1.16.0
tabulate==	0.9.0
tensorboard==	2.15.1
tensorboard-data-server==	0.7.2
tensorflow==	2.15.0
tensorflow-estimator==	2.15.0
tensorflow-intel==	2.15.0
tensorflow-io-gcs-filesystem==	0.31.0
termcolor==	2.4.0
threadpoolctl==	3.2.0
tifffile==	2023.12.9
tinyec==	0.4.0
tqdm==	4.66.1
typing-extensions==	4.9.0
tzdata==	2023.3
urllib3==	2.1.0
werkzeug==	3.0.1
wheel==	0.42.0
wordcloud==	1.9.3
wrapt==	1.14.1
zipp==	3.17.0
